Extract error handler and fix some level store problems
