+++ .gitignore
@@ -1,4 +1,4 @@
 node_modules
 yarn-error.log
 .env
-./db
+db/
