+++ src/main.ts
@@ -1,6 +1,6 @@
 import { Client, Intents, Message } from 'discord.js'
 import { fromEvent, Observable } from 'rxjs'
-import { init } from './config'
+import config from './config'
 import * as Event from './jevent/JEvent'
 import logger from './logger'
 import { handleMessage } from './messageHandler'
@@ -26,7 +26,7 @@ const main = async () => {
       })
   })
 
-  client.login(init().token)
+  client.login(config.token)
   logger.info('Done creating client!')
 }
